+++ src/React/EventLoop/LibEventLoop.php
@@ -10,7 +10,7 @@ class LibEventLoop implements LoopInterface
     private $callback;
 
     private $timers = array();
-    private $timersGc = array();
+    private $timersGc;
 
     private $events = array();
     private $flags = array();
@@ -21,23 +21,17 @@ class LibEventLoop implements LoopInterface
     {
         $this->base = event_base_new();
         $this->callback = $this->createLibeventCallback();
+        $this->timersGc = new \SplQueue();
+        $this->timersGc->setIteratorMode(\SplQueue::IT_MODE_DELETE);
     }
 
     protected function createLibeventCallback()
     {
-        $timersGc = &$this->timersGc;
         $readCallbacks = &$this->readCallbacks;
         $writeCallbacks = &$this->writeCallbacks;
 
-        return function ($stream, $flags, $loop) use (&$timersGc, &$readCallbacks, &$writeCallbacks) {
+        return function ($stream, $flags, $loop) use (&$readCallbacks, &$writeCallbacks) {
             $id = (int) $stream;
-
-            if ($timersGc) {
-                foreach ($timersGc as $signature => $resource) {
-                   event_free($resource);
-                   unset($timersGc[$signature]);
-                }
-            }
 
             try {
                 if (($flags & EV_READ) === EV_READ && isset($readCallbacks[$id])) {
@@ -177,13 +171,19 @@ class LibEventLoop implements LoopInterface
         );
 
         $timer->signature = spl_object_hash($timer);
+        $timersGc = $this->timersGc;
+        $callback = function () use ($timer, $timersGc) {
+            foreach ($timersGc as $resource) {
+                event_free($resource);
+            }
 
-        $callback = function () use ($timer) {
             if ($timer->cancelled === false) {
                 call_user_func($timer->callback, $timer->signature, $timer->loop);
 
                 if ($timer->periodic === true) {
                     event_add($timer->resource, $timer->interval);
+                } else {
+                    $this->cancelTimer($timer->signature);
                 }
             }
         };
@@ -214,7 +214,7 @@ class LibEventLoop implements LoopInterface
 
             $timer->cancelled = true;
             event_del($timer->resource);
-            $this->timersGc[$signature] = $timer->resource;
+            $this->timersGc->enqueue($timer->resource);
             unset($this->timers[$signature]);
         }
     }